requirements.txt
fastapi
uvicorn
pydantic
python-multipart
firebase-admin
google-generativeai
